refactor: dashboard
