+++ src/components/LoadingSpinner.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { Loader2 } from "lucide-react";
+
+export function LoadingSpinner({ message = "Loading..." }: { message?: string }) {
+  return (
+    <div className="flex h-full items-center justify-center p-8">
+      <div className="flex flex-col items-center gap-3">
+        <Loader2 className="h-6 w-6 animate-spin text-blue-500" />
+        <p className="text-sm text-gray-600">{message}</p>
+      </div>
+    </div>
+  );
+}
