+++ src/features/dashboard-admin/hooks/useKycData.ts
@@ -1,0 +1,172 @@
+"use client";
+
+import { useState, useEffect, useCallback } from "react";
+import { type PaginationState } from "@tanstack/react-table";
+import { toStartOfDayZ, toEndOfDayZ } from "../utils/datetime";
+import { Filters, getRowTs, toDisplayRow } from "../utils/kycHelpers";
+import { apiClient } from "../libs/apiClient";
+import type { KycRequestApi } from "../types/kyc";
+import type { Kycrequest } from "../components/column";
+
+export function useKycData(defaultValues?: Partial<Filters>) {
+  // Loading/Error
+  const [isLoading, setIsLoading] = useState(true);
+  const [isRefetching, setIsRefetching] = useState(false);
+  const [error, setError] = useState<Error | null>(null);
+
+  // Data
+  const [items, setItems] = useState<(Kycrequest & { __keys: string })[]>([]);
+  const [total, setTotal] = useState<number>(0);
+  const [nextPage, setNextPage] = useState<number>(1);
+
+  // Pagination
+  const [pagination, setPagination] = useState<PaginationState>({
+    pageIndex: 0,
+    pageSize: 10,
+  });
+
+  // Draft Filters (ค่าที่ user พิมพ์/เลือก)
+  const [draftFilters, setDraftFilters] = useState<Filters>({
+    q: defaultValues?.q ?? "",
+    status: defaultValues?.status ?? "all",
+    startDate: defaultValues?.startDate ?? "",
+    endDate: defaultValues?.endDate ?? "",
+  });
+
+  // Applied Filters (ค่าที่ apply แล้ว)
+  const [appliedFilters, setAppliedFilters] = useState<Filters>(draftFilters);
+
+  // helpers
+  const isEmail = (text: string) => /\S+@\S+\.\S+/.test(text);
+  const isTransactionNumber = (text: string) => /^\d{1,10}$/.test(text);
+  const isThaiText = (text: string) => /[\u0E00-\u0E7F]/.test(text);
+
+  const buildFilterQuery = useCallback(
+    (page: number, limit: number) => {
+      const { q, status, startDate, endDate } = appliedFilters;
+
+      const params: Record<string, string> = {
+        page: String(page),
+        limit: String(limit),
+      };
+      let specialSingleThai: string | null = null;
+
+      const qTrim = q.trim();
+      if (qTrim) {
+        if (isEmail(qTrim)) {
+          params.email = qTrim;
+        } else if (isTransactionNumber(qTrim)) {
+          params.correlationId = qTrim;
+        } else if (isThaiText(qTrim)) {
+          const parts = qTrim.split(/\s+/).filter(Boolean);
+          if (parts.length === 1) {
+            specialSingleThai = parts[0];
+          } else {
+            params.firstNameThai = parts[0];
+            params.lastNameThai = parts.slice(1).join(" ");
+          }
+        } else {
+          params.correlationId = qTrim;
+        }
+      }
+
+      if (status && status.toLowerCase() !== "all") {
+        params.status = status;
+      }
+
+      if (startDate) params.startDate = toStartOfDayZ(new Date(startDate))!;
+      if (endDate) params.endDate = toEndOfDayZ(new Date(endDate))!;
+
+      return { params, specialSingleThai };
+    },
+    [appliedFilters]
+  );
+
+  const fetchData = useCallback(
+    async (page: number = 1, append: boolean = false) => {
+      try {
+        if (!append) setIsRefetching(true);
+
+        const { params, specialSingleThai } = buildFilterQuery(
+          page,
+          Math.max(100, pagination.pageSize * 2)
+        );
+
+        let newItems: (Kycrequest & { __keys: string })[] = [];
+
+        if (specialSingleThai) {
+          const [firstRes, lastRes] = await Promise.all([
+            apiClient.get<{ data: { items: KycRequestApi[] } }>("/api/company", {
+              params: { ...params, firstNameThai: specialSingleThai },
+            }),
+            apiClient.get<{ data: { items: KycRequestApi[] } }>("/api/company", {
+              params: { ...params, lastNameThai: specialSingleThai },
+            }),
+          ]);
+          newItems = [
+            ...(firstRes.data?.data?.items ?? []),
+            ...(lastRes.data?.data?.items ?? []),
+          ].map(toDisplayRow);
+        } else {
+          const res = await apiClient.get<{ data: { items: KycRequestApi[] } }>(
+            "/api/company",
+            { params }
+          );
+          newItems = (res.data?.data?.items ?? []).map(toDisplayRow);
+        }
+
+        const sorted = newItems.sort((a, b) => getRowTs(b) - getRowTs(a));
+
+        if (append) {
+          setItems((prev) => {
+            const existingIds = new Set(prev.map((i) => i.transactionNo));
+            return [...prev, ...sorted.filter((i) => !existingIds.has(i.transactionNo))];
+          });
+        } else {
+          setItems(sorted);
+          setPagination((p) => ({ ...p, pageIndex: 0 }));
+        }
+
+        setTotal(sorted.length);
+        setNextPage(page + 1);
+        setError(null);
+      } catch (e) {
+        setError(e as Error);
+      } finally {
+        setIsLoading(false);
+        setIsRefetching(false);
+      }
+    },
+    [buildFilterQuery, pagination.pageSize]
+  );
+
+  useEffect(() => {
+    fetchData(1, false);
+  }, [appliedFilters]);
+
+  const apply = useCallback(() => {
+    setAppliedFilters(draftFilters);
+  }, [draftFilters]);
+
+  const clearAll = useCallback(() => {
+    const cleared: Filters = { q: "", status: "all", startDate: "", endDate: "" };
+    setDraftFilters(cleared);
+    setAppliedFilters(cleared);
+  }, []);
+
+  return {
+    isLoading,
+    isRefetching,
+    error,
+    items,
+    total,
+    pagination,
+    setPagination,
+    fetchData,
+    apply,
+    clearAll,
+    draftFilters,
+    setDraftFilters,
+    appliedFilters,
+  };
+}
